smartinbox_env/security/pgp.py
# smartinbox_env/security/pgp.py
# ─────────────────────────────────────────────────────────────────
#  PGP Handler  |  Signature verification + envelope encryption
#  Production: use python-gnupg or cryptography library
# ─────────────────────────────────────────────────────────────────

from __future__ import annotations

import hashlib
from dataclasses import dataclass
from typing import Dict, Optional


@dataclass
class PGPVerifyResult:
    valid: bool
    fingerprint: Optional[str]
    signer: Optional[str]
    error: Optional[str]


@dataclass
class PGPEncryptResult:
    success: bool
    ciphertext: Optional[bytes]
    error: Optional[str]


# Simulated trusted key store  (production: load from GPG keyring)
_TRUSTED_KEYS: Dict[str, str] = {
    "security@company.com":   "AABBCCDDEEFF0011",
    "legal@lawfirm.com":      "1122334455667788",
    "compliance@company.com": "DEADBEEFCAFE1234",
    "devops@company.com":     "FEEDFACEDEADBEEF",
}


class PGPHandler:
    """
    Handles PGP signature verification and message encryption.

    In simulation mode (no actual GPG binary):
    • verify()  — checks sender against trusted key store
    • encrypt() — returns HMAC-SHA256 placeholder ciphertext

    Production swap:
    ─────────────────
    import gnupg
    gpg = gnupg.GPG()
    result = gpg.verify(email['pgp_signature'])
    """

    def verify(self, email: Dict) -> PGPVerifyResult:
        sender = email.get("sender", "").lower()
        sig    = email.get("pgp_signature")

        # No signature present
        if not sig and sender not in _TRUSTED_KEYS:
            return PGPVerifyResult(valid=False, fingerprint=None, signer=None, error="No signature")

        # Trusted sender in keyring
        if sender in _TRUSTED_KEYS:
            return PGPVerifyResult(
                valid=True,
                fingerprint=_TRUSTED_KEYS[sender],
                signer=sender,
                error=None,
            )

        return PGPVerifyResult(valid=False, fingerprint=None, signer=sender, error="Key not in keyring")

    def encrypt(self, plaintext: str, recipient: str) -> PGPEncryptResult:
        """Simulate PGP encryption (replace with actual GPG encrypt in prod)."""
        try:
            key_material = _TRUSTED_KEYS.get(recipient, "UNKNOWN_KEY")
            mock_cipher = hashlib.sha256(
                f"{key_material}:{plaintext}".encode()
            ).digest()
            return PGPEncryptResult(success=True, ciphertext=mock_cipher, error=None)
        except Exception as e:
            return PGPEncryptResult(success=False, ciphertext=None, error=str(e))
