pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-smartinbox-pro"
version = "0.1.0"
description = "SmartInbox-Pro OpenEnv email triage environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "gymnasium==0.29.1",
    "numpy>=1.26.4",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.5",
    "python-gnupg>=0.5.2",
    "aioimaplib>=1.1.0",
    "scikit-learn>=1.4.0",
    "huggingface_hub>=0.23.0",
    "datasets>=2.19.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
]

[project.scripts]
server = "smartinbox_pro.server.app:main"

[tool.setuptools]
include-package-data = true
packages = [
    "smartinbox_pro",
    "smartinbox_pro.server",
    "smartinbox_env",
    "smartinbox_env.ai_engine",
    "smartinbox_env.db",
    "smartinbox_env.imap_smtp",
    "smartinbox_env.security",
]
package-dir = { "smartinbox_pro" = ".", "smartinbox_pro.server" = "server" }
